+++ src/main/java/uos/uos25/order/dto/response/ReturnsGetResponseDTO.java
@@ -1,12 +1,12 @@
-package uos.uos25.returns.dto.response;
+package uos.uos25.order.dto.response;
 
 import java.time.LocalDateTime;
 
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import uos.uos25.product.dto.ProductInfo;
-import uos.uos25.returns.entity.Returns;
+import uos.uos25.headQuarter.dto.ProductInfo;
+import uos.uos25.order.entity.Returns;
 import uos.uos25.shop.dto.ShopInfo;
 
 @Data
