+++ src/main/java/uos/uos25/order/repository/ReturnsRepository.java
@@ -1,11 +1,11 @@
-package uos.uos25.returns.repository;
+package uos.uos25.order.repository;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import uos.uos25.returns.entity.Returns;
+import uos.uos25.order.entity.Returns;
 
 @Repository
 public interface ReturnsRepository extends JpaRepository<Returns, Long> {
